+++ Source/ReactiveCocoa/Moya+ReactiveCocoa.swift
@@ -16,7 +16,7 @@ public class ReactiveCocoaMoyaProvider<Target where Target: MoyaTarget>: MoyaPro
     }
     
     /// Designated request-making method.
-    public func request(token: Target) -> SignalProducer<MoyaResponse, MoyaError> {
+    public func request(token: Target) -> SignalProducer<Response, MoyaError> {
         
         // Creates a producer that starts a request each time it's started.
         return SignalProducer { [weak self] observer, requestDisposable in
@@ -25,7 +25,7 @@ public class ReactiveCocoaMoyaProvider<Target where Target: MoyaTarget>: MoyaPro
                     observer.sendFailed(.Underlying(error))
                 } else {
                     if let data = data, let statusCode = statusCode {
-                        observer.sendNext(MoyaResponse(statusCode: statusCode, data: data, response: response))
+                        observer.sendNext(Response(statusCode: statusCode, data: data, response: response))
                     }
                     observer.sendCompleted()
                 }